전체적인 코드 내용 정리 완료

+++ action/deleteScheduleAction.jsp
@@ -23,22 +23,29 @@
     //전페이지에서 온 데이터에 대해서 인코딩 설정
     request.setCharacterEncoding("UTF-8");
     
-    //계정 id값 받아오기 
+    //이 페이지의 계정 id값 받아오기 
     String id = request.getParameter("id"); 
     
     //스케줄 정보 받아오기
-    String date = request.getParameter("date"); 
     String scheduleIdxString = request.getParameter("scheduleIdx"); 
+    String date = request.getParameter("date"); 
 
-    if (scheduleIdxString == null) {
-        out.println("<div>올바른 접근이 아닙니다.</div>");
+    if (scheduleIdxString == null || date == null) {
+        out.println("<div>올바르지 않은 접근입니다.</div>");
         return;
     }
     
+    //스케줄idx값 int형으로 변환
     int scheduleIdx = Integer.parseInt(scheduleIdxString);
 
     //세션값 받아줌
     int accountIdx = (Integer)session.getAttribute("accountIdx");
+
+    if (accountIdx == 0) {
+        out.println("<div>올바르지 않은 접근입니다.</div>");
+        return;
+    }
+
 
     Connection connect = null;
     PreparedStatement query = null;
@@ -47,9 +54,11 @@
         Class.forName("com.mysql.jdbc.Driver");
         connect = DriverManager.getConnection("jdbc:mysql://localhost/9weekhomework","stageus","1234");
 
-        String sql = "DELETE FROM schedule WHERE idx = ? ";
+        //세션의 accountIdx값과 schedule테이블의 accountIdx값이 같은경우에만 쿼리문 전송되도록
+        String sql = "DELETE FROM schedule WHERE idx = ? AND account_idx = ? ";
         query = connect.prepareStatement(sql);
         query.setInt(1,scheduleIdx);
+        query.setInt(2,accountIdx);
         
         query.executeUpdate();
     }
@@ -57,7 +66,6 @@
         out.println("<div>예상치 못한 오류가 발생했습니다.</div>");
         return;
     }
-    
 %>
 
 <!DOCTYPE html>
